+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # Version 0.1.0
+- Add sr_cleanup method
+- Improve instance init checking the parameters type
+
+# Version 0.1.0
 - First version of the project.